python_app/requirements.txt
customtkinter>=5.2.0
Pillow>=10.0.0
pystray>=0.19.0
tkvideoplayer>=1.4
opencv-python>=4.8.0
